refactor: restructure components folder and update imports
