+++ src/App.js
@@ -1,7 +1,7 @@
 
 import React, { useState } from "react";
 import { Routes, Route, Navigate } from "react-router-dom";
-import Navbar from "./Components/Navbar";
+import { Navbar } from "./Components";
 
 import HomeSplit from "./pages/HomeSplit";
 import PlanTrip from "./pages/PlanTrip";
