+++ src/Components/Planner/index.js
@@ -1,0 +1,2 @@
+export { default as Card } from './Card';
+export { default as Tours } from './Tours';
