+++ src/Components/index.js
@@ -1,0 +1,5 @@
+// src/Components/index.js
+export { default as Navbar }  from './Layout/Navbar';
+export { default as Card }    from './Planner/Card';
+export { default as Tours }   from './Planner/Tours';
+export { default as Refresh } from './ui/Refresh';
